Add files from `flask_app`
The first step of gluing these two projects together is copying the
relevent files from the other project into this repo. The file contents
are unmodified, although some relative locations have changed.

The main `user_profiles_ex01.py` file and the `jinja_templates/` folder
have both been moved in to the `app/` subfolder, so to follow the layout
of this project.

The `user_profiles_ex01.json` file remains in the root folder. This is
where the hardcoded path expects it, so the code can remain unchanged.
Plus it's adjacent to the app.db file, so it feels right.

This means that the profiles and messages portion of flask_app can now
be launched from this project using this command:

```sh
FLASK_APP=app/user_profiles_ex01.py flask run --host "::1"
```

Finally, it is important to note that all of the files in this commit
are Leo McGuire's work. They have been copied from the latest commit on
the master branch: 67a38f5916cdac3844b291c0f5295338a7545655