Add helper bash script for launching dev server
The official Flask docs recommend using the environment variable to
activate the developmemt mode on the built in server. (As opposed to
using the Config object or the .flaskenv file.)

Also, on my machine in particular the server can't seem to bind to
either localhost or 127.0.0.1, but it does bind to the IPv6 version,
which is "::1".

So to save on the long command to remember and type, I made this script.